+++ src/app/about/about.scss
@@ -38,7 +38,6 @@
   &__content-container {
     display: flex;
     align-items: flex-start;
-    justify-content: center;
     margin-top: 60px;
     gap: 80px;
