+++ src/app/about/page.tsx
@@ -1,20 +1,9 @@
-import TeamBioCard from '@/components/team-bio-card/team-bio-card';
 import { getTeamMemberData } from '@/utils/get-team-members-data/get-team-members-data';
-import { Box, Button, Typography } from '@mui/material';
-import Container from '@mui/material/Container';
-import { Metadata } from 'next';
-import Script from 'next/script';
-import React from 'react';
-import './about.scss';
+import AboutClient from './page.client';
 
-interface TeamMember {
-  name: string;
-  job_title: string;
-  bio?: string;
-  linkedin_link: string;
-  github_link?: string;
-  profile_image_link?: string;
-}
+import { Metadata } from 'next';
+
+import React from 'react';
 
 export const metadata: Metadata = {
   title: 'About Us | CodeForBC',
@@ -24,70 +13,5 @@ export const metadata: Metadata = {
 export default async function About() {
   const teamMemberData = await getTeamMemberData();
 
-  const teamMemberJson = teamMemberData.map((teamMember: TeamMember) => ({
-    '@context': 'https://schema.org',
-    '@type': 'Person',
-    name: teamMember.name,
-    jobTitle: teamMember.job_title,
-    description: teamMember.bio || '',
-    url: 'https://www.codeforbc.ca/about',
-    image: teamMember.profile_image_link
-      ? `https://www.codeforbc.ca${teamMember.profile_image_link}`
-      : '',
-    sameAs: [teamMember.linkedin_link, teamMember.github_link || ''],
-  }));
-
-  return (
-    <Container maxWidth="lg" className="about-page">
-      <Script id="jsonLD_team-member" type="application/ld+json">
-        {JSON.stringify(teamMemberJson)}
-      </Script>
-      <Box className="about-page__heading-container">
-        <Typography className="page-heading about-page__heading" variant="h1">
-          Meet Our Team
-        </Typography>
-        <Typography className="about-page__text" variant="body1">
-          We are volunteer developers, designers, project managers, engineers,
-          marketers, organizers, strategists, and citizens committed to our
-          vision, and driven by our mission.
-        </Typography>
-      </Box>
-      <Box className="about-page__content-container">
-        <Box className="about-page__filter">
-          <Button className="about-page__filter-button">All Members</Button>
-          <Button className="about-page__filter-button">Executive</Button>
-          <Button className="about-page__filter-button">Product/UIUX</Button>
-          <Button className="about-page__filter-button">Developers</Button>
-          <Button className="about-page__filter-button">
-            Business Development
-          </Button>
-          <Button className="about-page__filter-button">Data</Button>
-        </Box>
-        <Box className="about-page__card-container">
-          {teamMemberData.map((teamMember: TeamMember) => (
-            <TeamBioCard key={teamMember.name} data={teamMember} />
-          ))}
-        </Box>
-      </Box>
-      <Box className="join-us-cta-card">
-        <Box className="join-us-cta-card__text-container">
-          <Typography className="join-us-cta-card__title" variant="body1">
-            Want to be a part of the team?
-          </Typography>
-          <Typography className="join-us-cta-card__text" variant="body2">
-            Contribute to Open Source Projects for BC, from Anywhere, with
-            CodeForBC.
-          </Typography>
-        </Box>
-        <Button
-          size="large"
-          className="join-us-cta-card__button"
-          variant="contained"
-          href="https://tally.so/embed/3jMe1x"
-        >
-          Join Us
-        </Button>
-      </Box>
-    </Container>
-  );
+  return <AboutClient teamMemberData={teamMemberData} />;
 }
